Remove rem and pixel fallback, instead default to px
We don’t have control over the font size set on the :root element or on
the html element. govuk_template sets the font size of the html element
to 62.5%, if our code is to work with existing projects - we need to
ensure that our components don’t scale according to this root size
which we don’t control.