Remove pattern check from date component
Use input type=“number” to trigger a numeric keyboard.

The pattern attribute applies when the value of the type attribute is
text, search, tel, url, email, or password, otherwise it is ignored.

Also remove min and max, as these will not stop a user entering a
number manually outside of this range - this only works for the spinner
control buttons, which we aren’t displaying for this component.