Increase severity of sass linting rules to 'error'
The `scss:lint` task will only exit with an error state if errors are reported.
Currently we only throw warnings (level 1 severity) for all violations. For now
we are bumping everything to error (level 2) - we can review which rules should
be errors and which should be warnings if we find the config too severe now.

We have not opted in to any extra rules as part of this change, we have just
bumped the existing warnings to errors.