Refactor the capitaliseComponentName helper method
As the method is always used to create a macro name from a component name, this renames the function to better reflect what it does, and moves the prefixing with `govuk` into the function. It also simplifies the code by using map rather than iterating over the array and adds a clearer comment block to explain its purpose.