Fix radio checked state rendering on smaller screens
When on smaller viewports such as on a mobile device, there seems to be a rounding error
which results in you being able to see a small white dot on the checked radio input.

By setting the background colour to the `currentColor` we ensure if a round error happens
the background will look the same as the border, removing any possible rendering issue.