Remove use of font-size-adjust
Because of the way we're currently using the tabular numbers typeface, which only contains tabular numbers and 'falls through' to nta, applying font-size-adjust to text that uses the tabular numbers typeface causes the browser to made adjustments to NTA - which is not what we want.

Now that we're providing weight overrides (e.g. .govuk-!-w-bold) this is hard to manage, and may result in some very unpredictable behaviour.

Given the fact that font-size-adjust is currently only supported by Firefox, we're going to remove it for now. Ideally we will revisit this decision if we are able to move to supporting tabular numbers as a feature of a single typeface, and/or if browser support improves.