Update visually hidden helper, add focussable version
Adopt the latest version from HTML5 boilerplate[1] which adds declarations for:

- clip-path, as clip is deprecated and has been removed from the Web standards, but is still needed for older browsers
- white-space, to prevent text wrapping onto multiple lines off screen and read as if there was no whitespace between the words

Because clip path was not previously being used, this also adds it in the logical place in the sass linter config so that the linter does not insist on declaring it last.

This also adds a version which allows for the element to be focussed using the keyboard or assistive technologies which we can use for the skip link component. For this version I’ve removed padding and border declarations to make it easier to style the focussed element, but this may mean you need to do extra work to ensure the element is truly hidden when not focussed.

[1]: https://github.com/h5bp/html5-boilerplate/blob/76f6b7d95358ffabb6ac6c4bc8d44a5af35fea8d/dist/css/main.css#L128-L168