Remove font face and weight declarations
The font-face is being set twice in a lot of the components - once by the typography mixin and once in the component itself. This removes the duplication by removing all font-face declarations.

It also replaces font-weight declarations with the appropriate mixin, so that e.g if in the future we choose to re-introduce font-size-adjust we can ensure that it’s applied correctly according to the font-weight.

Finally, it moves all typography mixin includes to the top of their respective rulesets.