Create details marker using border-based triangles
This allows us to consistently style the marker across browsers and resolves the issue with the Firefox marker disappearing because of our use of `display: inline-block`. For more details, see the implementation.md file added as part of this commit.

We also need to use clip-path to ensure that if browser colours are overriden the marker still appears as a triangle rather than turning into a square.

This also correctly sets the text colour for the disclosed text and adjusts the spacing so that the disclosed text is aligned with the summary.