Correctly sort components and examples in lists
Because these are defined as an object, their order is undefined which means every time you refresh the page they appear in a slightly different order.

Nunjucks has a built in filter 'dictsort' to handle this sort of problem, so applying that to both the component and example lists fixes it.