Fix variable scope
Because we currently use the `postcss-nested` plugin when copying Sass to the packages folder, we end up removing the nesting here and so we end up with $offset being defined in a different ruleset to where it's being used, which means it is not part of the same scope and we get an undefined variable error.

We need to revisit whether it makes sense to use this plugin, and make sure that we can identify issues like this sooner (rather than waiting until we try and build new packages) but for now let's fix the immediate issue.