Apply core link styling via classes
Change the core link styling so that it is applied via classes, rather than styling a elements directly. This brings it into line with the way other core styling is applied, and enforces our principle that GOV.UK Frontend should not have opinions about basic HTML elements.

Introduce comments to try and explain what some of the less obvious code is doing.