Remove redundant 'reset' for print styles
The intention of this appears to be to ensure that javascript and anchor links are not suffixed with the URL when printed.

However, the two sets of selectors are mutually exclusive - if a link starts with `javascript:` or `#` then it cannot also start with `/`, 'http://` or `https://`, so the previous rule cannot possibly applied in order to need resetting.