Add focus state to components that relied on global link styling
These components previously relied on the global link styling to set the focus state, which is no longer part of frontend.

To avoid duplication of the focus state, add an additional focusable helper designed for text-based focusable elements which includes setting the background colour of the target element to the focus colour.