Allow HTML to be passed to error summary list
This component used to call the list component, but when that was removed it was updated with a simple loop over the items array.

This updates the loop to allow users to pass HTML for an entry in the list, and also ensures that if they do not provide an href the text or html does not get wrapped in a link.