Remove the link component
The link styles have been moved to the 'core' layer, which means that the only thing the link component is really providing now is a 'download link' variant.

The download link variant needs more design work (it is currently missing a lot of the functionality present in download links on GOV.UK, such as displaying file sizes and formats) and is not a component that currently exists in Elements or Frontend Toolkit, so is being removed until we can spend some proper time on it.

The link component will need to be removed from packages and dist when the next version of Frontend is released.