Update warning text README
In [#432](https://github.com/alphagov/govuk-frontend/pull/432) changes were made directly to README.md that should have been made in index.njk, mainly to update references to the component name, but also to change the 'description' to remove a reference to legal consequences.

This updates index.njk to include the change to the description and regenerates the readme with the correct whitespace and capitalisation as well as changes to the arguments table and to the macro example.