Include pseudo classes in review app build
This installs the postcss pseudo-classes plugin which automatically adds in ‘companion classes’ where pseudo-selectors are used. This allows us to add class names (:hover, :visited) to force the styling of a pseudo-selector, allowing us to easily preview these style states in the review app.

We only include it as part of the postcss step for non-production builds – i.e. when building for the review app, not for the dist folder.