Add examples of button in different states
Use the new pseudo class companion classes to simulate active, hover and focus states for the button.

This is more of a proof of concept than anything - there may be other states that need to be considered, and there are definitely other components for which it would be useful to have examples of different states.