Add ‘no visited state' link variant
Add an ‘no visited state' link variant which can be used where it is not helpful to distinguish between visited and non-visited links.

For example, navigation links to pages with dynamic content like admin dashboards. The content on the page is changing all the time, so the fact that you’ve visited it before is not important.

Closes #297