Use placeholders when extending in prose scope
This updates the prose scope (and the things the prose scope depends on) such that we only ever @​extend placeholder classes (classes starting with % which are not outputted into the CSS themselves)

This abstraction gives us control over which occurrences of the selector are included by the @​extend. Without this change, occurrences of e.g. the heading classes within the fieldset component are also ending up being ‘applied’ to the prose scope, leading to selectors like `.govuk-c-fieldset__legend .govuk-prose-scope h4` in the CSS. This makes no sense, as the prose scope should never be used within the fieldset legend.

With this change, only occurrences of the placeholder class are included when @​extend’ing. This means that we can define e.g. the typographic contextual adjustments at the placeholder level and these _will_ be applied to the prose scope, but other references to the concrete classes will have no effect on the prose scope.