Use new input border colour variable for borders
At the minute we are using the `$govuk-text-colour` variable for the colour of form control borders, which doesn’t make much sense semantically.

This introduces a new variable, `$govuk-input-border-colour` which is set to the same `$govuk-black` that `$govuk-text-colour` currently refers to, and updates form components to use it for the border colour.