Remove IE7(?) fix
This code was copied from Frontend Toolkit. Unfortunately, there’s nothing in the comment or its history to tell us what version of IE8 the issue it is trying to address occurs in.

As far as I can tell, this is not an issue that affects IE8 (I can’t see any difference to the button in any state in IE8 when this line is removed, and references to this `filter: chroma` hack tend to talk about IE7) so I think we can remove it.