Add check for any new component's package.json If a new component has been added to src it needs a package.json file when it gets copied to packages/
We use to do this manually and our tests that run after `build:packages`
task would fail.

With this script, when `npm run pre-release` is run, the script will
check if any new folders exist in packages after the copy task. If so,
it will create a sample package.json file with actual component name,
version 0.0.0 and latest version of globals as dependency.

It will also remind the developer to amend ALL/package.json file with
new dependencies.

When developer indicates they have done the above, the script will
and the rest of the `npm run pre-release` tasks will run.