Ensure component has a role of `contentinfo`
https://www.w3.org/TR/wai-aria-1.1/#contentinfo)

Note: This decision has been made given that most uses of this role tend to be used directly on a `<footer>` element.
Assumption made is that is most predictable for users of assistive technologies.

The spec indicates that `contentinfo` is useful for "Examples of information included in this region of the page are copyrights and links to privacy statements.", which may indicate that it might be better placed around the 'meta' section of this component.
    Should be challenged if user research indicates this is an issue.