Add bottom margin to error summary component
This adds a bottom margin to the error summary component.

The error summary should always be the first thing on the page, after any breadcrumbs or back link. This means it will always appear before a heading, in which case we know that we will always need to add a margin to separate it from the heading.