Fix inline radios
Introduce a modifier at the block level which makes all radios within that block inline. You can’t add classes to the individual radio items using the macro, and it doesn’t really make sense to require the modifier to be applied to every individual element.

Add an example of inline radios so that they are represented in the review app.

Update the readme.