Include component previews using absolute URLs
Without this change, if you view a component with a URL that includes a trailing slash, the browser will try and load the previews relative to that path, so e.g. /components/checkboxes/checkboxes/preview.

The chances of this happening are increased when developing locally on a port that might also be used to run the Design System, as the Design System appears to return 301 redirects from trailing-slash-less URLs to their slashed counterparts, and the browser will cache these redirects and apply them to Frontend. This can be hard to debug.