Filter out files from component list
Currently if you have e.g. a .DS_Store file in the src directory it is included in the expected list of components and your tests will fail.

Update the file helper to filter out anything that’s not a directory from this list, and make it clearer what the various exported lists represent.