Automatically rasterise media queries for IE8
Rather than relying on the user remembering to specify another setting when building a stylesheet that targets IE8, modify the media queries settings to use sensible defaults based on the value of `$govuk-is-ie8`.

This means we need to ‘promote’ the `$govuk-is-ie8` variable to the settings layer so that we can include it before the media queries settings.