Fix table captions to allow heading classes
Table CSS overrides Core CSS, which includes the heading classes.

This update changes the example to use the heading class to reproduce the issue.

Then updates the CSS to inherit from the top level table element.

This allows for the heading to not get overriden in child elements.

Originally raised by Adam Silver, thanks!