Remove redundant font-family rule in button
Remove explicit font-family definition from the button scss. We don't need it because we get it as part of the included `govuk-font-regular` mixin which defines the font for us.

This will remove a redundant duplicated font-family rule and also mean that the button will correctly fall back to the print font when printed.