Render button shadow as border in IE8
IE8 does not support box-shadow, so use a border instead.

For browsers other than IE8 we use a transparent border to support users who override their browser colours, but the intersection of IE8 users who override their browser colours is likely to be negligible, so it should be OK to override this.