Omit empty labels
This ensures that the label element is not outputted if no text or HTML are provided. This allows for other form control components to be used with a separate label, e.g. if a user wants to have an input and a button on one line with a separate label above.