Remove ‘shorthand’ font helpers
These are being removed in favour of using `govuk-font` throughout - this decouples the codebase from the typography scale (allowing it to be overridden by consumers in the future), reduces the number of mixins in our public API and gives us greater control over being able to e.g. warn if points in the scale are going to be deprecated.