Save changed package-lock
There seems to be multiple changes to `package-lock` at the moment
with multiple changes to package version numbers. This is despite
the fact that last time the `package.json` was changed, we also
committed the `package-lock` in the PR.

As we are doing a release of GOV.UK Frontend, going to commit this
into the repo.

We need to keep an eye on this though.