Remove README from assets folder
This duplicates content from /README.md which means there’s a risk of the two getting out of sync.

We also suggest that users should either copy the entire assets folder or mount it as part of their application, which would result in this README being part of their public facing website which is not desired behaviour.