Add support for customising item inputs' field names in date-input
Add support for a new `item.{}.fieldName` parameter on the `date-input`
macro. This allows the author to explicitly set the field name on each
item input rather than rely on the `<param.name>-<item.name>` default
convention, which is not always convenient and couples the field name to
the displayed label too tightly.

Also fix a few minor missing/incorrect parts in the date-input README.