Pin node version to prevent Heroku failing
In #827 we removed pinning teh node version,
instead specifiying the minimum version.

This resulted in Heroku failing as there is an error
in the gulp taks in node 10.5.0 Heroku installs.

Until we decide to upgrade to Gulp 4 we need to
pin the version in package.json.

Alternative option is to remove version entirely, although
we want to set that for anyone wanting to contribute using
`yarn` instead of `npm`.