Remove internal changelog entry requirement
The primary use case for the CHANGELOG is to make sure users can react to any new release.

For example if there are breaking changes, they need to be able to action those.

Internal changes do not impact the running of GOV.UK Frontend in our users applications.

This commit removes this requirement to document changes, so we do not have to require,
contributors to update the CHANGELOG for a trivial change.