Pin gulp-better-rollup to 3.1.0
gulp-better-rollup 3.2.0 bumps rollup from 0.56.5 to 0.60.7 – unfortunately rollup 0.60 drops support for the legacy flag and thus Internet Explorer 8.

Until we find ourselves needing any of the features of the newer versions, we’ll pin to the older version of rollup in order to maintain support for IE8 for as long as possible.