Fix phase banner being affected by global styles
When global styles are enabled, the phase banner uses the wrong font size.

This is because the font-size of 16px is currently set on the parent `.govuk-phase-banner`, but `p.govuk-phase-banner__content` doesn't have a `font-size` set, so inherits the font size from the global paragraph styling.

Moving the font-size declaration onto `.phase-banner__content` fixes this.