Avoid `display: inline-block` for header content areas.
`display: inline-block` is sensitive to whitespace, so if there are characters in between two divs
that will be output when they are rendered.

This means when columns are exactly 33.3% and 66.6% to sit alongside each other,
the whitespace that is output misaligns the columns forcing them onto a new line.

Using floats is similar to how our grid system works, but requires a clearfix on the parent container,
so that the component does not collapse.