Fix ‘backwards compatibility’ for the date input.
The previous implementation meant that if your input _does not_ have the name ‘year’ it would always get the class ‘govuk-input--width-2`.

Because this class is defined last [1] – after every other width class – this makes it impossible to make any field that does not have the name ‘year’ use any other width than 2 characters.

This commit addresses this by:

- not adding a width class if any other width class is present within the `item.classes` (it’s relatively a naive check for the string ‘govuk-input--width-‘ but it should do the job)
- only adding appropriate classes if the input has the name ‘day’, ‘month’ or ‘year’

It also adds tests for this behaviour.

Fixes #907

[1]: https://github.com/alphagov/govuk-frontend/blob/8f2fe567865729716fa4a04b0c6eeb5860da7441/src/components/input/_input.scss#L71-L73