Only underline logo on hover (with custom colours)
Because all borders will become text colour when a user overrides colour in their browser, the transparent border will become always visible.

By only applying the border on hover we avoid this which means we get the correct behaviour even when colours are customised. The layout is not affected because we offset the additional border with a negative margin.

Removing any border-color attribute means that we'll use the currentColor for the border - making this explicit results in IE8 ignoring the rule because it doesn't understand the currentColor variable.