Fix tests for character count with govuk-textarea
Fix tests by using `js-character-count` for tests as this
contains the functionality - although this
is different from how we do it in other components, this should
be looked at as part of the tests card.

The hint and error message tests are still failing as they
are "hijacked" by the character count - we'll look at this
as part of the hint card.