Fix examples that use Date Input component
Since the class name is no longer automatically added (as changed in #969), we have to either provide the width class ourselves or omit the `items` attribute in order to get the ‘default’ items array, which includes the classes.

Where possible, I’ve opted for the latter – the only case where this isn’t currently possible is the passport expiry date example in the ‘error summary with messages’ example, where I’ve added the width classes.