Only allow textareas to be resized vertically
Prevent the textarea component from being resized beyond the bounds of its container, by only allowing it to be resized vertically.

Potential issues for the user are that it would make it difficult or impossible to reduce the width should a click + drag take the textbox beyond the viewport or otherwise overflow hidden regions.

This also makes it less likely that resizing the textarea will cause other issues with the layout of the page.

Fixes #966