Remove anchor styling in govuk-lists
This need can be satisfied by enabling global styles, which is on by default in the GOV.UK Prototype Kit.

I have checked our [documentation](https://design-system.service.gov.uk/styles/typography/#lists) and we do not recommend our users do this.

With this both in mind this feels reasonable to remove this functionality as a breaking change.