Fix robots header to prevent indexing
Because the various routes in the review app do not call `next`, the middleware that sets the `X-Robots-Tag` header is currently never called. We could call `next()` from every route, but that seems fragile, so instead we can just move the middleware so that it is defined before the routes, which means it will be called.

Also, we can add tests for this. So we do.