Add support for attributes on table cells
It’s currently not possible to attach HTML attributes to table cells.
This PR makes it possible by using the same object pattern that is used
on other components (and also the <table> element).

I’ve updated the nunjucks template as well as the YAML and the README
to reflect the new functionality.