Replace background image with inline SVG
This swaps out the + and - icon from being a background-image SVG to an inline SVG.

The SVG itself is also simplified to consist of two rounded rectangles on top of each other to form a + icon (with the vertical bar hidden to form the – icon).

Having the SVG inline enables it to change colour when the page colours are overridden for accssibility reasons.