Fix footer display in IE8
At the moment footer looks broken in IE8. It doesn't like the padding
being set on top and bottom of the container, resulting in an increased
padding at the bottom and a "hanging" top border.

This work moves the responsive padding to the 'meta' element for IE8,
but keeps it on the wrapper for IE9 and modern browsers.