Add 5px bottom margin to bullet-less list items
On small screens list items have no bottom margin (with a 5px bottom margin being applied at tablet and above)

When the govuk-list class is applied on it's own without any modifier, the <li>'s appear to run into each other on mobile, as the spacing between the list items is the same as the spacing between each line.

(When a bullet or number modifier is added e.g. class="govuk-list govuk-list--bullet" the bullet or number indicates the start of a new list item, so the spacing does not need to be adjusted in that case.)

This change applies the spacing to the govuk-list class for all breakpoints, then adjusts the behaviour when bullet or number modifiers are added.

Co-authored-by: David House <david.house@digital.cabinet-office.gov.uk>