Deprecate govuk-grid-row mixin
Wrapping the row functionality up in a mixin offers little other than the ability to use a custom name, but we’re not sure there’s a need for users to do that.

Simplify the code by deprecating the mixin and just using the code to create the concrete class directly.