Fix bug of <h2> and <div> inside <h2>
Summary is now specified as an option in Nunjucks and is placed after
the heading in the markup. This achieved the most consistent results when testing with
assistive tech. On encountering the heading, the assistive tech read out
the following:
- NVDA (Firefox): clickable [heading text] button collapsed [summary text] heading level 2
- Jaws 16 and Jaws 2018 (IE11): [heading text] heading level 2 button collapsed [summary text]
- iOS VoiceOver (Safari): [heading text] button heading level 2 expanded [summary text] double tap
to collapse
- Mac Voiceover (Safari): [heading text] you are currently on a heading level 2

Mac VoiceOver doesn't give out a lot of information or read the summary and I can't find good
examples of this being done on collapsible elements so hoping someone else in the community might
have ideas and I'm also going to run it past the Accessibility team at GDS. An alternative
(where Mac Voiceover does read out the summary) is to nest the summary inside the <h2> but this
seriously limits what could be placed in the summary as <h2> can only contain
[flow content](https://html.spec.whatwg.org/multipage/dom.html#flow-content) so it seems
likely that this would lead to users inadvertently creating non-semantic markup.

Also add .govuk-accordion_header wrapper for all elements contained in a
header. Using `<header>` was also considered but NVDA reads out "banner landmark" (preceding the
above description it reads out) on encountering it which could be confusing more than
helpful since it banner landmarks should be top level page items. Reference:
https://www.w3.org/TR/wai-aria-practices/examples/landmarks/banner.html

This also fixes some styling problems resulting from moving things around in the markup.

The next commit does some standardisation of class names, partly resulting from the markup
changes in this commit.