Refactors Accordion.js
1. init method had a lot of code in it and so I moved most of it out into
separate methods.

So now its easy to see that it initialises the section headers & creates
the controls.

As part of this refactoring, I changed how "readState" method worked. When
init method was running it would iterate through the headers twice, once
to setup the markup for the sections headers and then again to read the
previous state and set them. Instead this would be better to do it in one
transaction so it only loops through the headings once.

2. refactor checkIfAllSectionsOpen method

There is no need to loop through all the sections to work out if they are
expanded. We can find the length of sections that are expanded against
the total number of sections and if they are equal then it means
all sections are open.

3. refactors storeState to stop iterating through all sections

Everytime a user clicked on a section header storeState would be called
against each section in the accordion and set state for each section. Its
not a issue when we only have a few sections but the more sections there
are the more unnecessary  work the browser needs to do. Now we only call
storeState on the section that was clicked.

5. Refactors SessionStorage support

We should cache the result of SessionStorage support instead of constantly
looking it up and setting/removing session data.  We should proberly
move the SessionStorage helper method out to its own file as it might
be useful for other components.