Expand focus state to fill the heading
...so that it matches the hover state

This change was requested by the Working Group.

If we continue to keep the focus styles (yellow outline) on
the button, it is not possible to include the summary in the
focus. Including the summary makes sense both visually and
on the basis of the fact that it is read out by screen
readers on encountering the button.

This solution might however be too much "magic" as we're
moving the focus styles away from where the focus actually is
(the button). Submitting this for review though, will discuss
with team/community.