Update date input component to render `autocomplete`
WCAG 2.1 introduces a new success criteria "1.3.5 Identify Input Purpose" which
"is to ensure that the purpose of a form input collecting information about the
user can be programmatically determined, so that user agents can extract and present
this purpose to users using different modalities".

Broadly speaking, we can help service teams meet this criteria by making it easier
to pass autocomplete attributes to components.

At the minute this can be done by using the attributes option, but making it part of the
component API promotes the idea that this should be used when it makes sense to do so.