Remove hover state for accordion sections on mobile
On iOS the hover state could be 'stuck' when the user is not interacting with the accordion component anymore.

One alternative approach would be to only add a hover state when the device has hover capability, I chose not to do this since this feature is techincally still a candidate recommendation which means it could in theory be removed if something happens while they test this new feature.

In the future we could simplify this code if it becomes a stable recommendation.