Prevent summary list from wrapping mid-word
[PR #1169](https://github.com/alphagov/govuk-frontend/pull/1169) added in some CSS:
- ``word-break: break-word;``, followed by
- ``word-break: break-all;``

``word-break: break-word;`` is the favoured behaviour, but isn't fully supported.
``word-break: break-all;`` is the fallback behaviour, it sometimes results in unfavourable wrapping mid-word.

The order for this CSS needs to be changed to, otherwise the less favourable `break-all` takes precedence.
```CSS
word-break: break-all;
word-break: break-word;
```

**caniuse** said:
> Chrome, Safari and other WebKit/Blink browsers also support the unofficial break-word value

The implication here being that it is not fully supported, so ``word-break: break-all;`` is still needed as a fallback.

Discussed with [Colin Rotherham](https://github.com/colinrotherham)