Ensure font is not loaded twice on slow networks
Prior to this change if you were loading GOV.UK Frontend alongisde
alphagov/govuk_template you would have two @font-face declarations that
tell the browser how to load the 'nta' font.

This is normally not an issue since as soon as the browser downloads a
version it won't download a second copy of the font.

However when a user is on a slow connection, the browser can try to
download the font from the alphagov/govuk_template declaration and
GOV.UK Frontend declaration.
Which means that the end user will download the font twice (and it's variants)
resulting in a much slower experience.

This is even more of a problem for users on a slow network since their
bandwidth may already be constrained.

To avoid this we check if the user has indicated they are using
GOV.UK Template by checking the relevant compatibility flag.

If they have set this flag, we avoid outputting the declarations with the
assumption that they will already have them, avoiding the chance of the font being downloaded twice.

This feature can also be disabled with a top level $govuk-include-default-font-face flag to ensure that our users have full control.