Move '/full-page-examples' to their own route file
I think this is a better approach because we are following the existing
pattern of using routes and not having to require a file in the middle
of app.js only for that file to then require other files to adds
post routes to the app. Also theres no need to be passing the app round
to other files.

We should now be following a Route/Controller pattern and abstracting the
Controller methods out into their own file. This modularity improves the
readablitiy of the code and how app is setup. Its also easier to test
Existing tests however have not been changed or update but will be
in a future commit.