Inline requires instead of assigning them to variables
At first I thought it would nice to be clear what routers we were using
at the top by declaring them altogether. It was also difficult at first
because there was a lot of router details in and amongst this file.

Looking at the current state I think we should just inline the requires
to the routers as you suggest because all the routes are now being called
relatively close together and we do not use those variables anywhere else.