Rename border-width-mobile to reflect how it's used
Despite its name, the variable isn't used on mobile.

Standardise the border width on error summary on mobile to use
the standard border width.

From @dashouse): "The 4px border is used in the conditional reveals
because it has to be an even number in order to be centred under the 40px
checkbox or radio. 5px would put it 1px too far left or right and it was
noticeable. The error summary change is probably not needed."

Fixes https://github.com/alphagov/govuk-frontend/issues/1240