Adding govuk-focusable-text-link mixin
This approach uses box-shadow, which has broadly good support across browser.

We have found that Edge and Internet Explorer produce rendering issues.

Edge 18 has quite problematic issues but this can be seen on any website using custom
focus styles. Edge 18 supports 'drop-shadow()' which can work but only on elements that are `diplay: inline`, with this in mind we think the complexity is not worth the cost to users for all browsers.

Edge 19+ will be using the Chromium rendering engine so this will not be an issue.

Edge 17 and below, and IE 9-11 have lesser issues that we have determined will not cause a barrier.

See: https://gist.github.com/nickcolley/3788eefd34e7f04391de07867a85018b for an example of how `drop-shadow()` could work.

Co-authored-by: Nick Colley <nick.colley@digital.cabinet-office.gov.uk>