Remove incorrectly overriden error summary focus style
When we updated the the focus styles for to have the same colour as the body text, we forgot to update the specific overrides we need when using GOV.UK Frontend with GOV.UK Template (legacy).

This meant that when you focused a link in the error summary the color was red instead of black, removing this block achieves the result we want.