Improve tabs heading spacing on mobile and no-js view
Split from https://github.com/alphagov/govuk-frontend/pull/1326

This amends the tab headings spacing on mobile and no-js view to be
consistent with how we present the navigation on the Design System:

Adjusts the padding/margin so that with the the list version of component
that’s used on mobile and no-js we use have govuk-spacing(2) as bottom margin,
consistent with what we do with on-page navigation in GOV.UK Design System when
it's displayed as a list. For the desktop/js-enabled version of component remove
bottom margin and use top and bottom padding instead.

Also adjusts bottom margin of list item so that it’s applied on no-js version
and add some more spacing under the tabs title.