fixes tabs keyboard navigation bug in IE8
In IE8, the browser could not find the next/previous tab because it does
not support `nextElementSibling` and `previousElementSibling` DOM traversal
methods. To fix it I applied a polyfill for it.

Once the browser could find the tab we then had to find the `firstChildElement`
(i.e the anchor element) to add/edit the various data attributes to show/hide
the tab panel. Again IE8 doesn't support it and instead of introducing another
polyfill I used `querySelector instead to look up the "a". I assumed the
first element would always be anchor for navigation purposes and also the
nunjucks template uses an anchor with the class name that I'm looking up.