Add two polyfills for nextElementSibling and previousElementSibling
I've added two new files polyfill files although they didn't come directly
from polyfill.io they were based off a PR that was merged in the library
but not included in the new polyfill-library repo.
I've added comments pointing to the original pull request for the
detection and for the polyfill.