Remove govuk-focusable mixin
We are removing this mixin as it's no longer useful.

Many of the focus states now include edge cases or logic that'd require more focused mixins or parameters passed to the current mixins.

We're removing this mixin but may introduce more explicit mixins in the future for things like inputs.