Move aria-label to nav element in header
HTML validator flagged this as a ‘possible misuse of aria-label.’

This pattern of applying aria-label to the <ul> was copied from GOV.UK Template.[1]

Best practise is to add the aria-label to the element with the navigation role[2], which in this case would be the `<nav>`.

When a label is used on a navigation element, the term ‘navigation’ should be omitted[3], as ‘the screen reader will read both the role and the contents of the label’, so I’ve also removed the word ’navigation’ from the label.

[1]: https://github.com/alphagov/govuk_template/blob/f164b2825b1479d7c7d4bdbec4ba9c8e171f504c/docs/usage.md#propositional-title-and-navigation
[2]: https://www.w3.org/TR/wai-aria-practices/examples/landmarks/navigation.html
[3]: https://developer.mozilla.org/en-US/docs/Web/Accessibility/ARIA/Roles/Navigation_Role#Associated_WAI-ARIA_Roles_States_and_Properties