Remove unused icon assets
Neither of these images are currently used:

`assets/images/icon-important.png` was added as part of the warning text component in 1ac78a0, but never used.

`assets/images/icon-arrow-left.png` was originally added as part of the back link in 246cab6. The back link was then removed in 3aa3d76 (leaving the asset unused) and reintroduced in 769f419 but without using images.